chore(weave): pr feedback - move url_safety to helpers/, log blocked urls
