+++ weave/trace_server/helpers/url_safety.py
